Move new extensions around
